Rename State to TopLevelState
And this patch moves it into top_level module.
